requirements.txt
numpy>=1.24
opencv-python>=4.8
pandas>=2.0
PyYAML>=6.0
